:sparkles: Basic MQTT functionality, CLI
This includes the basic MQTT functionality needed to communicate with
a broker and setup the necessary channels. It also comes with a basic
CLI that right now just connects to the broker and implements the
discovery feature.

The `mosquitto.config` is included for use with the `eclipes-mosquitto`
Docker image for development purposes. Use it something like this:

```
docker run -it -p 1883:1883 -p 9001:9001 \
  -v $PWD/mosquitto.config:/mosquitto/config/mosquitto.conf:ro \
  eclipse-mosquitto
```