:art: :pencil: Refactor Device
Remove a number of methods from `Device` that dealt with getting/setting
the topic and publishing or subscribing to a feature. Instead attach
those methods to `Feature` so they can be called directly on them. This
feels much more natural and simplifies the API a bit as we can get rid
of some error checking (since you can't call those methods on features
that don't exist which we had to guard on before).

However, since the `Set` and `Watch` methods on features do require access
to the device transport I now have to maintain a reference to the partent
struct/device. An alternative approach could be to embed a reference to
the transport in the features too.

This also gets rid of the `GetTopic` and `SetTopic` methods entirely as
they're no longer necessary. Since the manager always sets the
`GetTopic` and `SetTopic` attributes on a feature they can be used
directly instead of going through the Device.