:ambulance: :pencil: Resubscribe on device update
When a device is updated we should resubscribe to the feature's topics.
If we do not then after a reconnect to the MQTT broker we no longer have
any subscriptions.

We can blindly resubscribe as the Eclipse MQTT library we use (Paho)
only allows one callback per topic. If one exists it'll replace it with
the new one, if not it'll add one.

Closes #9